Fix black formatting in merger core module
Fix quote style inconsistency detected by CI/CD black checks:
- Change single quotes to double quotes in core.py:1324-1325
- Ensures compliance with black==23.12.1 formatting standards

This resolves CI/CD quality check failures.

+++ enahopy/merger/core.py
@@ -1321,8 +1321,8 @@ class ENAHOGeoMerger:
                 self.logger.debug(
                     f"Normalizing integer dtypes for '{column}': {type1} and {type2} → int64"
                 )
-                df1[column] = df1[column].astype('int64')
-                df2[column] = df2[column].astype('int64')
+                df1[column] = df1[column].astype("int64")
+                df2[column] = df2[column].astype("int64")
                 return
 
         if type1 != type2:
